requirements.txt
grpcio
grpcio-tools
protobuf
pytest-grpc
